Simple turret enemy
Enemy that turns towards the player and shoots, as long as the player is in line of sight.
Added a mine that detects players entering an area.

+++ Assets/Code/Weapon Scripts/Bullet.cs
@@ -7,7 +7,7 @@ public class Bullet : MonoBehaviour
     public float speed = 50;
     public float lifetime = 10;
 
-    public float damage = 5;
+    public int damage = 5;
 
     Vector3 currentPos;
